starting to separate ocs and byzcoin
